大幅升级整体架构，还存在一些问题，逐步修正

+++ Z-Bishop/src/main/java/com/isahl/chess/bishop/protocol/mqtt/model/CodeMqtt.java
@@ -73,13 +73,13 @@ public enum CodeMqtt
     SUBSCRIPTION_IDENTIFIERS_UNSUPPORTED(-1, 0xA1),
     WILDCARD_SUBSCRIPTION_UNSUPPORTED(-1, 0xA2);
 
-    private final byte _V3Code;
-    private final byte _V5Code;
+    private final int _V3Code;
+    private final int _V5Code;
 
     CodeMqtt(int v3, int v5)
     {
-        _V3Code = (byte) v3;
-        _V5Code = (byte) v5;
+        _V3Code = v3;
+        _V5Code = v5;
     }
 
     @Override
